Update 2010-07-12-vignettes-with-css3-box-shadows.md
Hello!

I am suggesting to modify the code above to remove the "body:after" since this causes buttons and other attributes to lose functionality. I also added in the "margin: 0%" to ensure that the image would still fill the screen. 

It looks just as good as before, now it doesn't interfere with website functionality.

Thanks!